resolve moderate severity vulnerability (postCSS line return parsing error) separate components
