+++ KPI-team1-frontend/src/pages/KpiPage.tsx
@@ -9,6 +9,8 @@ import {
 import { supabase } from "../supabase";
 import { Kpi, KpiExtended, KpiValue } from "../model/kpi";
 import { getWeek, getQuarter } from "date-fns";
+import KpiDetailModalPage from "./KpiDetailModalPage";
+import { getDisplayValueByPeriodicity } from "../helpers/kpiHelpers";
 
 const HEADER_KPI_COLUMNS: GridColDef[] = [
   {
@@ -44,27 +46,12 @@ const HEADER_KPI_COLUMNS: GridColDef[] = [
     headerAlign: "center",
     align: "center",
     renderCell: params => {
-      if (params.value === null) return null;
       const periodicity = params.row.periodicity;
-      const date = new Date(params.value as string);
-      if (periodicity === "daily") {
-        return date.toLocaleDateString("en-CH");
-      } else if (periodicity === "quarterly") {
-        return `Q${getQuarter(date)}`;
-      } else if (periodicity === "yearly") {
-        return date.toLocaleDateString("en-CH", {
-          year: "numeric",
-        } as Intl.DateTimeFormatOptions);
-      } else if (periodicity === "weekly") {
-        return `Week ${getWeek(date, {
-          weekStartsOn: 1,
-          firstWeekContainsDate: 4,
-        })}`;
+      const date = params.value ? new Date(params.value as string) : null;
+      if (date === null || periodicity === undefined) {
+        return null;
       } else {
-        return date.toLocaleDateString("en-CH", {
-          month: "short",
-          year: "numeric",
-        } as Intl.DateTimeFormatOptions);
+        return getDisplayValueByPeriodicity(periodicity, date);
       }
     },
   },
@@ -141,205 +128,6 @@ export default function KpiPage(): JSX.Element {
     fetchKpiDefinitions();
   }, []);
 
-  const renderModalContent = () => {
-    const [selectView, setSelectView] = useState<"values" | "history">(
-      "values"
-    );
-    const [kpiValues, setKpiValues] = useState<KpiValue[]>([]);
-
-    const fetchKpiValues = async () => {
-      try {
-        if (selectedKpi) {
-          let { data: kpi_values, error } = await supabase
-            .from("kpi_values_period_standardized")
-            .select("*")
-            .eq("kpi_id", selectedKpi.kpi_id)
-            .eq("circle_id", currentCircle);
-
-          if (error) {
-            throw error;
-          }
-          setKpiValues(kpi_values || []);
-          console.log(kpi_values, "kpi_values");
-        }
-      } catch (error: any) {
-        alert(error.message);
-      }
-    };
-
-    useEffect(() => {
-      fetchKpiValues();
-    }, [selectedKpi]);
-
-    const renderAddNewValue = () => {
-      return (
-        <>
-          <div className="text-2xl ">
-            Set a new value for{" "}
-            <span className="font-medium">{selectedKpi?.kpi_name}</span>
-          </div>
-          <div className="flex justify-between my-2">
-            <label className="font-medium w-full mr-2">
-              Set a date
-              <input
-                className="block w-full p-2 border rounded-md"
-                name="set date"
-                type="date"
-              />
-            </label>
-            <label className="font-medium w-full">
-              Enter a new value
-              <input
-                className="block w-full p-2 border rounded-md"
-                name="new value"
-                type="number"
-              />
-            </label>
-          </div>
-          <div className="pt-4 flex justify-end">
-            <button className="w-28 h-10 mr-4 bg-white rounded border border-customYellow justify-center items-center gap-2 inline-flex text-zinc-700 text-base font-medium">
-              Cancel
-            </button>
-            <button className="w-28 h-10 bg-customYellow rounded justify-center items-center gap-2 inline-flex text-base font-medium">
-              Save
-            </button>
-          </div>
-        </>
-      );
-    };
-
-    const renderPreviousValues = () => {
-      const PREVIOUS_VALUES_COLUMNS: GridColDef[] = [
-        {
-          headerName: "Date",
-          field: "standardized_date",
-          flex: 1,
-          sortable: true,
-          headerAlign: "center",
-          align: "center",
-          renderCell: params => {
-            if (params.value === null) return null;
-            const periodicity = params.row.kpi_periodicity;
-            const date = new Date(params.value as string);
-            if (periodicity === "daily") {
-              return date.toLocaleDateString("en-CH");
-            } else if (periodicity === "quarterly") {
-              return `Q${getQuarter(date)}`;
-            } else if (periodicity === "yearly") {
-              return date.toLocaleDateString("en-CH", {
-                year: "numeric",
-              } as Intl.DateTimeFormatOptions);
-            } else if (periodicity === "weekly") {
-              return `Week ${getWeek(date, {
-                weekStartsOn: 1,
-                firstWeekContainsDate: 4,
-              })}`;
-            } else {
-              return date.toLocaleDateString("en-CH", {
-                month: "short",
-                year: "numeric",
-              } as Intl.DateTimeFormatOptions);
-            }
-          },
-        },
-        {
-          headerName: "Value",
-          field: "value",
-          flex: 1,
-          sortable: true,
-          headerAlign: "center",
-          align: "center",
-        },
-        {
-          headerName: "Status",
-          field: "comment",
-          flex: 1,
-          sortable: false,
-          filterable: false,
-          headerAlign: "center",
-          align: "center",
-        },
-      ];
-
-      return (
-        <>
-          <div className="mt-4 text-2xl">Previous Values</div>
-          <div className="">
-            <DataGrid
-              getRowId={row => row.kpi_value_history_id}
-              rows={kpiValues}
-              rowSelection={false}
-              columns={PREVIOUS_VALUES_COLUMNS}
-              classes={{
-                columnHeaders: "bg-customPurple",
-                columnHeader: "uppercase",
-              }}
-            />
-          </div>
-        </>
-      );
-    };
-
-    const renderSetTargetValue = () => {
-      return (
-        <>
-          <div className="mt-4 text-2xl">Set a target value for this year</div>
-          <div className="flex justify-between my-2">
-            <label className="font-medium w-full mr-2">
-              Due Date
-              <input
-                className="block w-full font-normal text-neutral-400 p-2 border rounded-md border-neutral-400"
-                name="target date"
-                type="date"
-              />
-            </label>
-            <label className="font-medium w-full">
-              Enter the target value
-              <input
-                className="block w-full font-normal text-neutral-400 p-2 border rounded-md border-neutral-400"
-                name="target value"
-                type="number"
-                placeholder="What's your target"
-              />
-            </label>
-          </div>
-        </>
-      );
-    };
-    return (
-      <>
-        <div className="text-center text-2xl font-light">
-          KPI - {selectedKpi?.kpi_name}
-        </div>
-        <div className="mb-2">
-          <button
-            className={`${
-              selectView === "values" ? "border-black" : " dark:text-gray-600"
-            } w-1/3 h-10 border-b-2 justify-center items-center font-medium`}
-            onClick={() => setSelectView("values")}
-          >
-            Values
-          </button>
-          <button
-            className={`${
-              selectView === "history" ? "border-black" : " dark:text-gray-600"
-            } w-1/3 h-10  border-b-2 justify-center items-center font-medium`}
-            onClick={() => setSelectView("history")}
-          >
-            History
-          </button>
-        </div>
-        {selectView === "values" ? (
-          <>
-            <div className="bg-gray-100 p-4">{renderAddNewValue()}</div>
-            {renderPreviousValues()}
-            {renderSetTargetValue()}
-          </>
-        ) : null}
-      </>
-    );
-  };
-
   const renderDataGrid = (periodicity: string) => {
     const filteredKpiDefinitions = kpiDefinitions.filter(
       item => item.periodicity === periodicity
@@ -378,10 +166,14 @@ export default function KpiPage(): JSX.Element {
 
   return (
     <>
-      <ModalRightSide isOpen={modalIsOpen} onRequestClose={handleOpenModal}>
-        {renderModalContent()}
-      </ModalRightSide>
-
+      {selectedKpi && (
+        <KpiDetailModalPage
+          isOpen={modalIsOpen}
+          onRequestClose={handleOpenModal}
+          kpi={selectedKpi}
+          circleId={currentCircle}
+        />
+      )}
       <div className="flex m-12">
         <div className="w-11/12 md:w-3/4  xl:w-800">
           <div className="text-2xl py-4 my-2 border-b border-gray-300">
